+++ src/HartreeFock.cpp
@@ -49,7 +49,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 	Potential.GenerateTrial(Orbitals);
 
 	// TODO I feel like this many for loops is a sin, rejigging how inputs work or vectorising would help -S.P.
-	for (int i = 0; i < Orbitals.size(); i++)
+	for (int i = 0; i < static_cast<int>(Orbitals.size()); i++)
 	{
 		N_elec_n0 = -1; // Electron does not screen itself //Orbitals[i].occupancy() - 1;
 		N_elec_n1 = 0;
@@ -57,10 +57,10 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 		Orbitals[i].Energy = 0;
 		shell_occupancies = Orbitals[i].get_subshell_occupancies(); // 
 		// Iterate through each subshell
-		for (int L = 0; L < shell_occupancies.size(); L++)
+		for (int L = 0; L < static_cast<int>(shell_occupancies.size()); L++)
 		{
 			// Get the screening contribution on an electron in the i'th orbital/shell with orbital L, due to the j'th orbital/shell.
-			for (int j = 0; j < Orbitals.size(); j++)
+			for (int j = 0; j < static_cast<int>(Orbitals.size()); j++)
 			{
 				if (Orbitals[j].occupancy() == 0) continue; // No contribution to screening. 
 				screening_shell_occupancies = Orbitals[j].get_subshell_occupancies();
@@ -79,7 +79,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 				}
 				// [nd] and [nf] groups closer to the nucleus.
 				else
-				{ 	for(int L_other = 0; L_other < screening_shell_occupancies.size(); L_other++){ 
+				{ 	for(int L_other = 0; L_other < static_cast<int>(screening_shell_occupancies.size()); L_other++){ 
 						if (L > L_other && Orbitals[i].N() >= Orbitals[j].N())
 						{ N_elec_n2_plus += Orbitals[j].occupancy();
 						}
@@ -102,7 +102,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 		}
 	}
 	// Quick and dirty approximation of shells to a p orbital. TODO
-	for (int i = 0; i < Orbitals.size();i++){
+	for (int i = 0; i < static_cast<int>(Orbitals.size());i++){
 		if(Orbitals[i].L() == -10){ 
 			Orbitals[i].set_L(1,false);  
 		}	
@@ -136,7 +136,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 
 // Check if the atom has a single (occupied) orbital
 	int num_occupied_orbs = 0, single_orb_idx = -1;
-	for (int i = 0; i < Orbitals.size(); i++)
+	for (int i = 0; i < static_cast<int>(Orbitals.size()); i++)
 	{
 		if (Orbitals[i].occupancy() != 0)
 		{
@@ -150,7 +150,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 
 	if (num_occupied_orbs == 1 && Orbitals[single_orb_idx].occupancy() == 1) Potential.Reset();
 	
-	for (int i = 0; i < Orbitals.size(); i++) {
+	for (int i = 0; i < static_cast<int>(Orbitals.size()); i++) {
 		if (std::isnan(Orbitals[Orbitals.size()-1].F[i])){throw std::runtime_error("F invalid pre-Master!");}
 		Master(&Lattice, &Orbitals[i], &Potential, Master_tolerance, log);
 		if (std::isnan(Orbitals[Orbitals.size()-1].F[i])){throw std::runtime_error("F invalid post-Master!");}
@@ -214,7 +214,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 		// Hartree-Fock loop without exchange. Local exchange is evaluated here.
 		while (E_max_error > LDA_tolerance || m < 2)
 		{
-			if (m > 20 * Orbitals.size()) {
+			if (m > 20 * static_cast<int>(Orbitals.size())) {
 				log << "Starting approximation does not converge... " << endl;
 				log.flush();
 				break;
@@ -234,7 +234,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 				}
 			}*/
 
-			for (int i = 0; i < Orbitals.size(); i++)
+			for (int i = 0; i < static_cast<int>(Orbitals.size()); i++)
 			{
 				/*if (i == single) {
 					E_rel_change[i] = 0;
@@ -251,7 +251,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 					log << endl;
 					Orbitals[i] = Orbitals_old[i];
 					E_max_error = 0;
-					for (int j = 0; j < Orbitals.size(); j++) {
+					for (int j = 0; j < static_cast<int>(Orbitals.size()); j++) {
 						if (E_rel_change[j] > E_max_error && j != i) E_max_error = E_rel_change[j];
 					}
 					Orbitals[i].Energy *= 1 + 0.99*E_max_error;
@@ -267,7 +267,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 
 			// Find new largest relative error.
 			E_max_error = 0;
-			for (int i = 0; i < Orbitals.size(); i++) {
+			for (int i = 0; i < static_cast<int>(Orbitals.size()); i++) {
 				if (Orbitals[i].Energy != Orbitals_old[i].Energy) {
 					E_rel_change[i] = fabs(Orbitals[i].Energy / Orbitals_old[i].Energy - 1);
 				}
@@ -305,7 +305,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 			{
 				if (m > max_HF_iterations) { break; }
 
-				for (int i = 0; i < Orbitals.size(); i++)
+				for (int i = 0; i < static_cast<int>(Orbitals.size()); i++)
 				{
 					if (i == single_orb_idx) continue;
 					if (E_rel_change[i] < E_max_error && m != 0) continue;
@@ -340,7 +340,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 						if (max_iter > 20)
 						{
 							int new_max = 0;
-							for (int j = 0; j < E_rel_change.size(); j++)
+							for (int j = 0; j < static_cast<int>(E_rel_change.size()); j++)
 							{
 								if (j == i) continue;
 								if (E_rel_change[new_max] < E_rel_change[j]) new_max = j;
@@ -369,7 +369,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 					} while (correction_scaling != 1);
 				}
 				E_max_error = 0;
-				for (int i = 0; i < Orbitals.size(); i++) {
+				for (int i = 0; i < static_cast<int>(Orbitals.size()); i++) {
 					// Find new largest relative error.
 					/*if (Orbitals[i].Energy != Orbitals_old[i].Energy) {
 						E_rel_change[i] = fabs(Orbitals[i].Energy / Orbitals_old[i].Energy - 1);
@@ -381,7 +381,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 				m++;
 				if (E_max_error < HF_tolerance && !Final_Check)
 				{
-					for (int i = 0; i < Orbitals.size(); i++)
+					for (int i = 0; i < static_cast<int>(Orbitals.size()); i++)
 					{
 						E_rel_change[i] = 1;
 					}
@@ -396,7 +396,7 @@ HartreeFock::HartreeFock(Grid &Lattice, vector<RadialWF> &Orbitals, Potential &P
 	if (m >= max_HF_iterations && log.is_open()) {
 		log << "====================================================================" << endl;
 		log << "Too many iterations in Hartree-Fock" << endl;
-		for (int i = 0; i < Orbitals.size(); i++) {
+		for (int i = 0; i < static_cast<int>(Orbitals.size()); i++) {
 			log << i + 1 << ") n = " << Orbitals[i].N() << " l = " << Orbitals[i].L()
 				<< " Energy = " << Orbitals[i].Energy << " Occup = " << Orbitals[i].occupancy() << endl;
 		}
@@ -412,9 +412,9 @@ double HartreeFock::OrthogonalityTest(vector<RadialWF> &Orbitals)
 	double ort = 0;
 	vector<double> density(lattice->size(), 0);
 
-	for (int i = 0; i < Orbitals.size(); i++)
+	for (int i = 0; i < static_cast<int>(Orbitals.size()); i++)
 	{
-		for (int j = i + 1; j < Orbitals.size(); j++)
+		for (int j = i + 1; j < static_cast<int>(Orbitals.size()); j++)
 		{
 			if (Orbitals[i].L() != Orbitals[j].L()) continue;
 			int max_R = max(Orbitals[i].pract_infinity(), Orbitals[j].pract_infinity());
@@ -443,14 +443,14 @@ int SetBoundaryValues(Grid* Lattice, RadialWF* Psi, Potential* U)
 	Psi->F[0] = Correction*pow(Lattice->R(0), (Psi->L() + 1));
 	Psi->G[0] = Correction*pow(Lattice->R(0), Psi->L())*(Psi->L() + 1 + U->V[0] * Lattice->R(0)*Lattice->R(0) / (Psi->L() + 1));
 
-	unsigned int Turn = 1;
-	unsigned int infinity = Lattice->size() - 1;
+	int Turn = 1;
+	int infinity = static_cast<int>(Lattice->size()) - 1;
 	double S = 0, P = 0;
 
 	if (Psi->Energy < 0.)
 	{
 
-		while (Psi->Energy > U->V[Turn] && Turn < Lattice->size() - 1)
+		while (Psi->Energy > U->V[Turn] && Turn < static_cast<int>(Lattice->size()) - 1)
 		{
 			Turn++;
 		}
@@ -462,7 +462,7 @@ int SetBoundaryValues(Grid* Lattice, RadialWF* Psi, Potential* U)
 		{
 			S += sqrt(U->V[infinity] - Psi->Energy)*Lattice->dR(infinity);
 			infinity++;
-			if (infinity >= (Lattice->size() - 1)) { break; }
+			if (infinity >= (static_cast<int>(Lattice->size()) - 1)) { break; }
 		}
 
 		//set boundary values for inwards integration
@@ -523,7 +523,7 @@ int SetBoundaryValuesApprox(Grid * Lattice, RadialWF * Psi, Potential* U)
 
 		a[0] = 1;
 		b[0] = -lambda;
-		for (int i = 1; i < a.size(); i++)
+		for (int i = 1; i < static_cast<int>(a.size()); i++)
 		{
 			a[i] = a[i - 1] * (Psi->L()*(Psi->L() + 1) - (sigma - i)*(sigma - i + 1)) / (2 * i*lambda);
 			b[i] = a[i - 1] * ((sigma + i)*(sigma - i + 1) - Psi->L()*(Psi->L() + 1)) / (2 * i);
@@ -939,8 +939,8 @@ double HartreeFock::Conf_En(vector<RadialWF> &Orbitals, Potential &U)
 	double wght = 0; // Complete shell occupancy.
 	double eAB = 0;
 	double angular = 0;
-	for (int a = 0; a < Orbitals.size(); a++) {
-		for (int b = a; b < Orbitals.size(); b++) {
+	for (int a = 0; a < static_cast<int>(Orbitals.size()); a++) {
+		for (int b = a; b < static_cast<int>(Orbitals.size()); b++) {
 			if (b != a) wght = 1.*Orbitals[b].occupancy();
 			else wght = 0.5*(4*Orbitals[b].L() + 2)*(Orbitals[b].occupancy() - 1)/(4*Orbitals[b].L() + 1);
 			if (wght == 0) continue;
@@ -969,13 +969,13 @@ double HartreeFock::Conf_En(vector<RadialWF> &Orbitals, vector<RadialWF> &Virtua
 	int OneElec = 0;
 
 	vector<RadialWF*> Occupied(0);
-	for (int i = 0; i < Orbitals.size(); i++) {
+	for (int i = 0; i < static_cast<int>(Orbitals.size()); i++) {
 		if (Orbitals[i].occupancy() != 0) {
 			Occupied.push_back(&Orbitals[i]);
 			OneElec += Orbitals[i].occupancy();
 		}
 	}
-	for (int i = 0; i < Virtual.size(); i++) {
+	for (int i = 0; i < static_cast<int>(Virtual.size()); i++) {
 		if (Virtual[i].occupancy() != 0) {
 			Occupied.push_back(&Virtual[i]);
 			OneElec += Orbitals[i].occupancy();
@@ -988,8 +988,8 @@ double HartreeFock::Conf_En(vector<RadialWF> &Orbitals, vector<RadialWF> &Virtua
 	double wght = 0; // Complete shell occupancy.
 	double eAB = 0;
 	double angular = 0;
-	for (int a = 0; a < Occupied.size(); a++) {
-		for (int b = a; b < Occupied.size(); b++) {
+	for (int a = 0; a < static_cast<int>(Occupied.size()); a++) {
+		for (int b = a; b < static_cast<int>(Occupied.size()); b++) {
 			wght = Occupied[a]->occupancy();
 			if (b != a) wght *= Occupied[b]->occupancy();
 			else wght *= 0.5*(4*Occupied[b]->L() + 2)*(Occupied[b]->occupancy() - 1)/(4*Occupied[b]->L() + 1);
